+++ lib/lazy_builder.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
 
 typedef LazyOnErrorWidgetBuilder = Widget Function(
@@ -6,6 +8,8 @@ typedef LazyOnErrorWidgetBuilder = Widget Function(
 ///Can be used to build a [Widget] lazily, such that, we can get the [RenderBox]
 ///for the [Widget] then build another widget accordingly.
 class LazyBuilder extends StatefulWidget {
+  ///This can be used to build a [Widget] lazily, such that, we can get the
+  ///[RenderBox] for the [Widget] then build another widget accordingly.
   const LazyBuilder({
     Key? key,
     required this.child,
@@ -13,6 +17,8 @@ class LazyBuilder extends StatefulWidget {
     this.onError,
     this.forceTry,
     this.emptyOnError = false,
+    this.constraints,
+    this.preventRedundantRebuilds = false,
   }) : super(key: key);
 
   ///[Widget] for which a [RenderBox] is generated.
@@ -43,20 +49,32 @@ class LazyBuilder extends StatefulWidget {
   ///[Container].
   final bool emptyOnError;
 
+  ///If is not null, then the [child] is displayed with these [BoxConstraints]
+  ///as parent's [BoxConstraints] in offstage. i.e [child]'s [RenderBox] is
+  ///found under these [BoxConstraints].
+  final BoxConstraints? constraints;
+
+  ///If it is true, then the [RenderBox] will be generated again, only if the
+  ///[child]'s key has changed, instead of every build.
+  final bool preventRedundantRebuilds;
+
   @override
   State<LazyBuilder> createState() => _LazyBuilderState();
 }
 
 class _LazyBuilderState extends State<LazyBuilder> {
-  final GlobalKey key = GlobalKey();
-  late final bool forceTry;
+  late GlobalKey key;
+  late bool forceTry;
+  OverlayEntry? overlayEntry;
   var offstage = true;
   Object? error;
   RenderBox? renderBox;
 
-  @override
-  void initState() {
+  void _init() {
+    log('initiating', name: '$this');
+    offstage = true;
     forceTry = widget.forceTry ?? false;
+    key = GlobalKey();
 
     WidgetsBinding.instance.addPostFrameCallback((_) {
       try {
@@ -64,10 +82,47 @@ class _LazyBuilderState extends State<LazyBuilder> {
       } catch (err) {
         error = err;
       }
-      if (forceTry || renderBox != null) offstage = false;
+      if (forceTry || renderBox != null) {
+        offstage = false;
+      }
       setState(() {});
     });
+  }
+
+  @override
+  void initState() {
+    _init();
     super.initState();
+  }
+
+  @override
+  void didUpdateWidget(covariant LazyBuilder oldWidget) {
+    if (widget.preventRedundantRebuilds) {
+      if (widget.child.key != oldWidget.child.key) _init();
+    } else {
+      if (widget.child != oldWidget.child) _init();
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+
+  Widget get offstageChild {
+    final child = Container(
+      key: key,
+      child: widget.child,
+    );
+    final parentConstraints = widget.constraints;
+    return Offstage(
+      offstage: true,
+      child: parentConstraints == null
+          ? child
+          : OverflowBox(
+        maxHeight: parentConstraints.maxHeight,
+        minHeight: parentConstraints.minHeight,
+        maxWidth: parentConstraints.maxWidth,
+        minWidth: parentConstraints.minWidth,
+        child: child,
+      ),
+    );
   }
 
   @override
@@ -78,15 +133,7 @@ class _LazyBuilderState extends State<LazyBuilder> {
         (widget.emptyOnError ? emptyContainerFn : defaultChildFn);
     final builder = widget.builder ?? defaultChildFn;
 
-    if (offstage) {
-      return Offstage(
-        offstage: true,
-        child: Container(
-          key: key,
-          child: widget.child,
-        ),
-      );
-    }
+    if (offstage) return offstageChild;
     if (renderBox == null) {
       return onError(context, error, widget.child);
     }